backend/requirements.txt
Flask
Flask-CORS
mysql-connector-python
pandas
scikit-learn
statsmodels
python-dotenv